requirements.txt
-f /usr/share/pip-wheels
alembic==1.4.0
Babel==2.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
psutil==5.7.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
redislite==5.0.165407
requests==2.23.0
rq==1.3.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
Whoosh==2.7.4
WTForms==2.2.1
